Updating <<-EOS.undent to <<~EOS
I received an error of Calling <<-EOS.undent is disabled after running brew install cachalot. The error suggested using <<~EOS instead and making a PR for the issue. After a little research this seems to be a known issue and fix. Attached are a couple links I found useful. I hope this helps!

Thanks!

https://github.com/Homebrew/homebrew-php/issues/4727
https://anti-pattern.com/strip-leading-whitespace-from-heredocs-in-ruby